refactor: use typescript for non-api modules
